feat(deps): bumps jsdom to latest version
