+++ .github/workflows/deploy-staging.yml
@@ -1,0 +1,45 @@
+name: staging
+
+on:
+  workflow_dispatch:
+
+permissions: read-all
+
+jobs:
+  deploy_dev:
+    name: Preparing Production release
+    if: ${{ github.repository == 'sws2apps/jw-epub-parser' && github.ref == 'refs/heads/beta' }}
+    environment:
+      name: Staging
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        node-version: [16.x]
+
+    steps:
+      - name: Checkout for release preparation
+        uses: actions/checkout@ec3a7ce113134d7a93b817d10a8272cb61118579
+        with:
+          ref: beta
+          persist-credentials: false
+
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@1f8c6b94b26d0feae1e387ca63ccbdc44d27b561
+        with:
+          node-version: ${{ matrix.node-version }}
+
+      - name: Install package dependencies
+        run: npm ci
+
+      - name: Build for release
+        run: npm run build
+
+      - name: Semantic Release
+        id: semantic
+        uses: cycjimmy/semantic-release-action@5982a02995853159735cb838992248c4f0f16166
+        with:
+          semantic_version: 19.0.2
+        env:
+          GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
