--- .github/workflows/release.yaml
@@ -1,65 +1,0 @@
-name: JW EPUB Parser Analyze & Publish to NPM
-
-on:
-  push:
-    branches: [ main, beta, alpha ]
-  pull_request:
-    branches: [ main, beta, alpha ]
-
-jobs:
-  analyze:
-    name: Code QL Analysis
-    runs-on: ubuntu-latest
-    permissions:
-      actions: read
-      contents: read
-      security-events: write
-
-    strategy:
-      fail-fast: false
-      matrix:
-        language: [ 'javascript' ]
-
-    steps:
-    - name: Checkout repository for code analysis
-      uses: actions/checkout@v2
-
-    - name: Initialize CodeQL
-      uses: github/codeql-action/init@v1
-      with:
-        languages: ${{ matrix.language }}
-
-    - name: Perform CodeQL Analysis
-      uses: github/codeql-action/analyze@v1
-    
-  publish:
-    name: Version & Publish
-    if: ${{ github.repository == 'sws2apps/jw-epub-parser' && (github.ref == 'refs/heads/main' || github.ref == 'refs/heads/beta' || github.ref == 'refs/heads/alpha') }}
-    runs-on: ubuntu-latest
-    needs: [analyze]
-
-    strategy:
-      matrix:
-        node-version: [16.x]
-
-    steps:
-    - name: Checkout repository for npm publish
-      uses: actions/checkout@v2
-
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v2
-      with:
-        node-version: ${{ matrix.node-version }}
-        cache: 'npm'
-
-    - name: Install dependency
-      run: npm ci
-
-    - name: Run testing
-      run: npm run test
-
-    - name: Semantic release & Publish to npm
-      run: npm run semantic-release
-      env:
-        NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
