+++ .github/workflows/staging-to-dev.yml
@@ -1,0 +1,30 @@
+name: staging-to-development
+
+on:
+  workflow_dispatch:
+
+permissions: read-all
+
+jobs:
+  stg_to_prod:
+    name: Rebase from staging to development
+    environment:
+      name: Development
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout for rebase preparation
+        uses: actions/checkout@ec3a7ce113134d7a93b817d10a8272cb61118579
+        with:
+          ref: alpha
+          token: ${{ secrets.GH_TOKEN }}
+
+      - name: Do git rebase
+        run: |
+          git fetch origin
+
+          git checkout alpha
+
+          git rebase origin/beta
+
+          git push -f origin alpha
