+++ .github/workflows/staging-to-prod.yml
@@ -1,0 +1,30 @@
+name: staging-to-production
+
+on:
+  workflow_dispatch:
+
+permissions: read-all
+
+jobs:
+  stg_to_prod:
+    name: Rebase from staging to production
+    environment:
+      name: Production
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout for rebase preparation
+        uses: actions/checkout@ec3a7ce113134d7a93b817d10a8272cb61118579
+        with:
+          ref: main
+          token: ${{ secrets.GH_TOKEN }}
+
+      - name: Do git rebase
+        run: |
+          git fetch origin
+
+          git checkout main
+
+          git rebase origin/beta
+
+          git push -f origin main
