+++ .npmignore
@@ -4,3 +4,4 @@ test
 .github
 
 release.config.js
+rollup.config.js
